Reorganize monero-rpc-odoo code
